runner: split off local-runner into its own file
